+++ packages/smarthr-ui/src/intl/DateFormatter.tsx
@@ -2,9 +2,9 @@
 
 import dayjs from 'dayjs'
 
-import { useIntl } from './useIntl'
+import { useDateFormat } from './useDateFormat'
 
-import type { FormatDateProps } from './useIntl'
+import type { FormatDateProps } from './useDateFormat'
 
 /**
  * 日付を現在のロケールに応じてフォーマットして表示するコンポーネント
@@ -56,6 +56,6 @@ export const DateFormatter = ({
   date: string | Date
 }) => {
   const date = dayjs(orgDate).toDate()
-  const { formatDate } = useIntl()
+  const { formatDate } = useDateFormat()
   return <time dateTime={date.toISOString()}>{formatDate({ ...rest, date })}</time>
 }
