+++ packages/smarthr-ui/src/intl/TimeFormatter.tsx
@@ -3,9 +3,9 @@
 import dayjs from 'dayjs'
 import { memo } from 'react'
 
-import { useIntl } from './useIntl'
+import { useDateFormat } from './useDateFormat'
 
-import type { FormatTimeProps } from './useIntl'
+import type { FormatTimeProps } from './useDateFormat'
 
 /**
  * 時刻を現在のロケールに応じてフォーマットして表示するコンポーネント
@@ -36,7 +36,7 @@ import type { FormatTimeProps } from './useIntl'
 export const TimeFormatter = memo<Omit<FormatTimeProps, 'date'> & { date: string | Date }>(
   ({ date: orgDate, ...rest }) => {
     const date = dayjs(orgDate).toDate()
-    const { formatTime } = useIntl()
+    const { formatTime } = useDateFormat()
     return <time dateTime={date.toISOString()}>{formatTime({ ...rest, date })}</time>
   },
 )
