+++ packages/smarthr-ui/src/intl/index.ts
@@ -1,5 +1,6 @@
 export { IntlProvider } from './IntlProvider'
 export { useIntl } from './useIntl'
+export { useDateFormat } from './useDateFormat'
 export { Localizer } from './Localizer'
 export { DateFormatter } from './DateFormatter'
 export { TimeFormatter } from './TimeFormatter'
